--- a/.github/dagger.json
+++ b/.github/dagger.json
@@ -1,6 +1,6 @@
 {
   "name": "ci",
-  "engineVersion": "v0.16.2",
+  "engineVersion": "v0.16.3",
   "sdk": {
     "source": "go"
   },