--- a/.github/main.go
+++ b/.github/main.go
@@ -11,7 +11,8 @@ import (
 const (
 	daggerVersion      = "v0.16.3"
 	upstreamRepository = "dagger/dagger"
-	defaultRunner      = "ubuntu-latest"
+	ubuntuVersion      = "24.04"
+	defaultRunner      = "ubuntu-" + ubuntuVersion
 	publicToken        = "dag_dagger_sBIv6DsjNerWvTqt2bSFeigBUqWxp9bhh3ONSSgeFnw"
 	timeoutMinutes     = 20
 )
@@ -177,10 +178,9 @@ func (ci *CI) withTestWorkflows(runner *dagger.Gha, name string) *CI {
 		WithJob(runner.Job("scan-engine", "engine scan")).
 		With(splitTests(runner, "testdev-", true, []testSplit{
 			{"cgroupsv2", []string{"TestProvision", "TestTelemetry"}, &dagger.GhaJobOpts{
-				// HACK: our main runners don't support cgroupsv2
-				Runner: []string{"ubuntu-latest"},
-				// NOTE: needed to silence redis warning logs in tests
-				SetupCommands: []string{"sudo sysctl -w vm.overcommit_memory=1"},
+				// NOTE: Our CI runners do not support cgroupsv2 as of 2025.03
+				// @gerhard @matipan @jedevc have more details
+				Runner: []string{AltGoldRunner()},
 			}},
 			{"modules", []string{"TestModule"}, &dagger.GhaJobOpts{
 				Runner: []string{PlatinumRunner(true)},
@@ -301,6 +301,27 @@ func Runner(
 	)
 }
 
+// Assemble an alternative runner name for a workflow
+func AltRunner(
+	cpus int,
+) string {
+	mem := cpus * 2
+	runner := fmt.Sprintf(
+		"nscloud-ubuntu-%s-amd64-%dx%d",
+		ubuntuVersion,
+		cpus,
+		mem)
+
+	// Fall back to default runner if repository is not upstream
+	// (this is GHA DSL and will be evaluated by the GHA runner)
+	return fmt.Sprintf(
+		"${{ github.repository == '%s' && '%s' || '%s' }}",
+		upstreamRepository,
+		runner,
+		defaultRunner,
+	)
+}
+
 // Bronze runner: Multi-tenant instance, 4 cpu
 func BronzeRunner(
 	// Enable docker-in-docker
@@ -335,4 +356,9 @@ func PlatinumRunner(
 	dind bool,
 ) string {
 	return Runner(3, daggerVersion, 32, true, dind)
+}
+
+// Alternative Gold runner: Single-tenant with Docker, 16 cpu
+func AltGoldRunner() string {
+	return AltRunner(16)
 }