--- a/.github/workflows/engine-cli.gen.yml
+++ b/.github/workflows/engine-cli.gen.yml
@@ -19,7 +19,7 @@ concurrency:
 jobs:
     cli-test-publish:
         runs-on:
-            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-16c-st' || 'ubuntu-latest' }}
+            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-16c-st' || 'ubuntu-24.04' }}
         name: cli-test-publish
         steps:
             - name: Checkout
@@ -216,7 +216,7 @@ jobs:
         timeout-minutes: 20
     engine-lint:
         runs-on:
-            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-16c-st' || 'ubuntu-latest' }}
+            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-16c-st' || 'ubuntu-24.04' }}
         name: engine-lint
         steps:
             - name: Checkout
@@ -413,7 +413,7 @@ jobs:
         timeout-minutes: 20
     engine-test-publish:
         runs-on:
-            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-16c-st' || 'ubuntu-latest' }}
+            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-16c-st' || 'ubuntu-24.04' }}
         name: engine-test-publish
         steps:
             - name: Checkout
@@ -610,7 +610,7 @@ jobs:
         timeout-minutes: 20
     scan-engine:
         runs-on:
-            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-4c' || 'ubuntu-latest' }}
+            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-4c' || 'ubuntu-24.04' }}
         name: scan-engine
         steps:
             - name: Checkout
@@ -807,7 +807,7 @@ jobs:
         timeout-minutes: 20
     scripts:
         runs-on:
-            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-4c' || 'ubuntu-latest' }}
+            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-4c' || 'ubuntu-24.04' }}
         name: scripts
         steps:
             - name: Checkout
@@ -1004,7 +1004,7 @@ jobs:
         timeout-minutes: 20
     testdev-cgroupsv-2:
         runs-on:
-            - ubuntu-latest
+            - ${{ github.repository == 'dagger/dagger' && 'nscloud-ubuntu-24.04-amd64-16x32' || 'ubuntu-24.04' }}
         name: testdev-cgroupsv2
         steps:
             - name: Checkout
@@ -1101,9 +1101,6 @@ jobs:
                 # Run a simple query to "warm up" the engine
                 dagger version
                 dagger core version
-              shell: bash
-            - name: sudo sysctl -w vm.overcommit_memory=1
-              run: sudo sysctl -w vm.overcommit_memory=1
               shell: bash
             - name: scripts/exec.sh
               id: exec
@@ -1265,7 +1262,7 @@ jobs:
         timeout-minutes: 30
     testdev-cli-engine:
         runs-on:
-            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-32c-dind-st' || 'ubuntu-latest' }}
+            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-32c-dind-st' || 'ubuntu-24.04' }}
         name: testdev-cli-engine
         steps:
             - name: Checkout
@@ -1523,7 +1520,7 @@ jobs:
         timeout-minutes: 30
     testdev-container:
         runs-on:
-            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-32c-dind-st' || 'ubuntu-latest' }}
+            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-32c-dind-st' || 'ubuntu-24.04' }}
         name: testdev-container
         steps:
             - name: Checkout
@@ -1781,7 +1778,7 @@ jobs:
         timeout-minutes: 30
     testdev-everything-else:
         runs-on:
-            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-32c-dind-st' || 'ubuntu-latest' }}
+            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-32c-dind-st' || 'ubuntu-24.04' }}
         name: testdev-everything-else
         steps:
             - name: Checkout
@@ -2039,7 +2036,7 @@ jobs:
         timeout-minutes: 30
     testdev-module-runtimes:
         runs-on:
-            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-32c-dind-st' || 'ubuntu-latest' }}
+            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-32c-dind-st' || 'ubuntu-24.04' }}
         name: testdev-module-runtimes
         steps:
             - name: Checkout
@@ -2297,7 +2294,7 @@ jobs:
         timeout-minutes: 30
     testdev-modules:
         runs-on:
-            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-32c-dind-st' || 'ubuntu-latest' }}
+            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-32c-dind-st' || 'ubuntu-24.04' }}
         name: testdev-modules
         steps:
             - name: Checkout