--- a/.github/workflows/github.gen.yml
+++ b/.github/workflows/github.gen.yml
@@ -19,7 +19,7 @@ concurrency:
 jobs:
     github:
         runs-on:
-            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-4c' || 'ubuntu-latest' }}
+            - ${{ github.repository == 'dagger/dagger' && 'dagger-g3-v0-16-3-4c' || 'ubuntu-24.04' }}
         name: Github
         steps:
             - name: Checkout